docker_fastapi/code/requirements.txt
#  Flask==2.3.2
pymongo==4.4.1
uvicorn==0.27.1
asgiref==3.8.0
python-dotenv==1.0.1
fastapi==0.110.3
streaming-form-data==1.15.0
pydantic==2.7.1
pydantic_settings==2.2.1
